+++ js/jquery.terminal.d.ts
@@ -195,6 +195,11 @@ declare namespace JQueryTerminal {
         escape: boolean;
     };
 
+    type promptOptions = {
+        delay?: number;
+        typing?: boolean;
+    };
+
     type InterpreterItem = {
         completion: "settings" | JQueryTerminal.Completion;
         history?: boolean;
@@ -239,6 +244,7 @@ declare namespace JQueryTerminal {
         invokeMethods: boolean;
         allowedAttributes: Array<RegExp | string>;
         delay: number;
+        ansi: boolean;
         typing: boolean;
         flush: boolean;
         formatters: boolean;
@@ -254,6 +260,7 @@ declare namespace JQueryTerminal {
         invokeMethods?: boolean;
         delay?: number;
         typing?: boolean;
+        ansi?: boolean;
         allowedAttributes?: Array<RegExp | string>;
         unmount?: JQueryTerminal.EchoEventFunction;
         onClear?: JQueryTerminal.EchoEventFunction;
@@ -651,7 +658,7 @@ interface JQueryTerminal<TElement = HTMLElement> extends JQuery<TElement> {
     set_position(pos: number, relative?: boolean): JQueryTerminal;
     get_position(): number;
     insert(str: string, stay?: boolean): JQueryTerminal;
-    set_prompt(prompt: JQueryTerminal.ExtendedPrompt): JQueryTerminal;
+    set_prompt(prompt: JQueryTerminal.ExtendedPrompt, options?: JQueryTerminal.promptOptions): JQueryTerminal | JQuery.Promise<void>;
     get_prompt<T extends JQueryTerminal.ExtendedPrompt>(): T;
     set_mask(toggle?: boolean | string): JQueryTerminal;
     get_output<T extends JQueryTerminal.Lines | string[]>(raw?: boolean): T;