+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.27.0
+### Feature
+* add CSS and option `ansi` to echo to properly render ANSI art
+* add typing animation to set_prompt [#673](https://github.com/jcubic/jquery.terminal/issues/673)
+* return promise from `set_prompt` and `echo` when using typing animation
+* new emoji
+### Bugfix
+* fix bug in cache (when echo same string two times in a row) [#672](https://github.com/jcubic/jquery.terminal/issues/672)
+* fix exception when calling term::echo with typing animation
+* fix regression in width ([chess demo](https://codepen.io/jcubic/pen/NWxOZQE))
+* fix usage in NodeJS (regression since 2.22.0)
+
 ## 2.26.0
 ### Feature
 * Better API for prompt typing animation with term::read and term::echo